Move AuthorityRequirement tests to separate package
