+++ kie-dmn/kie-dmn-validation/src/test/resources/org/kie/dmn/validation/authorityrequirement/AUTHREQ_DEP_REQ_AUTH_NOT_KNOWLEDGESOURCE.dmn
