+++ kie-dmn/kie-dmn-validation/src/test/resources/org/kie/dmn/validation/authorityrequirement/AUTHREQ_DEP_REQ_DEC_NOT_DECISION.dmn
