+++ kie-dmn/kie-dmn-validation/src/test/resources/org/kie/dmn/validation/authorityrequirement/AUTHREQ_MISSING_DEPENDENCY_REQ_DEC.dmn
